✨ add npm structs. Prepare module to integrate in dependency module
